requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
httpx==0.24.1
websockets==10.4
websocket-client==1.6.3
web3==6.11.2
xrpl-py==2.2.0
prometheus-client==0.17.1
pydantic==2.4.2
python-dotenv==1.0.0
redis==5.0.1
finnhub-python==2.4.19
sentry-sdk==1.35.0

torch==2.8.0
numpy==1.26.0
aiohttp==3.10.5
